Full compatibility for Linux Windows and MacOS
Now we are using prebuilt binary cython dll.
So we dont need to compile them runtime anymore.
The only problem, is it is version dependant.
For now, it is only working on python 3.5.x
And not compatible with 3.6 (or 3.4).
We need to rebuild each cython on the correct OS for the correct build.
It is also working only on x64 system